Moved line iterator from DIMACSReader to external class.
